minor changes, renames, styles
